VC5Decompressor: HighPassBand::decode(): refactor, -3%
Instead of having a single flat loop that assumes that the output data
has no padding between rows, which is true now, but not a feature,
do a 2D loop over each output pixel, and fill it.
IMO this is a better structure anyways.

This measures as -3% perf improvement both in single-thread
and and 32-thread usage..

+++ src/librawspeed/decompressors/VC5Decompressor.h
@@ -212,7 +212,8 @@ class VC5Decompressor final : public AbstractDecompressor {
   };
   std::vector<ReconstructionStep> reconstructionSteps;
 
-  static inline void getRLV(BitPumpMSB* bits, int* value, unsigned int* count);
+  static inline void getRLV(BitPumpMSB* bits, int16_t* value,
+                            unsigned int* count);
 
   void parseLargeCodeblock(const ByteStream& bs);
